A simplified and more rational version of frontend
Location search page has been removed and location search had been merged with the search filters in the Search page. The separate restaurant window, appearing over the Search page after clicking on the restaurant block in the list, has been removed, as well as the Menu button and separate Menu window. Instead, the restaurant block in the search page is expected to maximize as it is being clicked on. It will then reveal the rest of the restaurant information, including Daily Menus if they are available