Code optimization
The filtering procedures have been split into function which are created outside of the route function. Thus, filtering processes by cuisines, distance and dietary options are performed by 3 separate function respectively, which are called by each other in the route function. This has been done in order to not overload the route function and to structure the code