Adding function for likes and unlikes
Function like_and_unlike is created outside of the route and is called in the route function in case the request method is POST when the like button is clicked. The function receives id of the restaurant from the like button of this restaurant's card. In case this id is together with the current user id in the restaurant_user junction table, it means the restaurant already has been liked and should be instead unliked, so it is deleted from the junction table. Otherwise, it is being liked and added to the table.

The check_liked_restaurants function is returning the list of liked restaurants of the current user to the template where it is used to identify the state of the like button.

The main route has been renamed to /home for easier identification