Changing the like button
There is created a condition for the like button based on the list of liked resturants of the current user, passed to the template from the /home route. If the id of the current restaurant is present in this list, the like button appearance is set to a filled star, while if it is not, the star is empty. The use of corresponding images is implemented with the help of svg's and sprite

In addition, at the end of the html file the favourites.js file is connected