Adding function for liking restaurants
The like_and_unlike function is created outside of the the route and is called inside the route function in case the request method is POST so the button is clicked. Through the POST method the restaurant id is passed from the clicked like button to the function. It is then checked whether this restaurant has been liked - whether this restaurant's id is together with the current user's id in the restaurant_user table. If it is, then the click on the button leads to deleting the restaurant from favourites, otherwise, it is added to the junction table with favourites. 

The check_liked_restaurants function returns a list of liked restaurants' ids of the current user. This list is then passed to the template where it is used to define the like button appearance based on the presence of the restaurant in this list.

In addition, the route name has been changed to /home for easier identification