Like button variations
The like button appearance depends on the conditions. In case the restaurant id is in the list of liked restaurants' ids, the button is given the appearance of a filled star, otherwise - an empty star. For this purpose, svg and sprite are used

At the bottom of the file, favourites.js is connected for the button functionality