Making conditions for displaying data
There have been created conditions for showing the rows from the restaurant data set.
The html structure for distance and time data is only created under the condition this data is present in the dataset. In case the users have not specified their location, nothing appears then.

The same result is implemented for dietary options. In this case, HTML depends on values of these fileds and HTML structure is only initiated in case of positive value - 1.

The phone numbers fields as well are shown only in case there values are not 0

For credit_card and takeaway fields HTML structure is  always created but its contents depend on the negative or positive value of these fields.

In addition, in case the user has applied different filters for the combination which no restaurants hav been found, a text informing about this outcome is created in the place of restaurant cards