Adding signup page template
The signup template includes 4 inputs for the user's name, email, password and its confirmation. In addition there is a link to the login page in case the user already has registered. When the password is being entered, it is shadowed. It also has a set minimum and maximum length: 8 -20 characters. The email is automatically validated by checking if it includes @ and characters before and after this character