Adding signup route
In the route function user input from the signup form is received. For name and email fields escape_string is applied in order to escape special characters and preserve the database from sql injections. The user input is validated first throw checking if the password and confirmed password match. If they don't a message about it is send to the user. Otherwise, validation is taken to the next step - checking if the email is already in the database. If it is, again, the user receives a message about it in the page. If there is no such email in the database, the user can be added to it. For this case user password is hashed and then all credentials are added to the database. The user then gets a message about successful regestration.