Adding login route
In the route function users' credentials are received from the login form. escape_string is applied to email to avoid sql injections. First, it is checked if such email exists in the database. If it does not, the user receives an error message in the login page. Otherwise, the entered password is validated an compared with the one from the database - in the same record with the varified email. If passwords do not match - the user receives an error message. Otherwise, an access token is generated. It is linked to this user by setting the user's id from the database as identity. The token is set in the cookie which is set in the response object. Response is then send to the browser where the cookie with the token appears. In the jwt configuration in the beginning of the file the location for sending this cookie has been set to /home route which is supposed to be a protected roite in the application and only accessable with jwt token from cookies. Cookies here are by default httpOnly. They can also be configured to be secure, but this confirguration can only be set to True in production mode for HTTPS connection. Here, in development mode, this configuration is now set to False. The life time of JWT has been set to 2 hours, after this period the token will expire. Cookies will only be available untill the browser is closed.