Adding jwt loader functions
If the access token is not valid for any reason (missing, expired, tampered with, etc) there is returned json in the format of {‘msg’: ‘why accessing endpoint failed’} along with an appropriate http status code (generally 401 or 422). With the jwt_manager loader functions it is possible to custmize this response.

expired_token_loader() is a function to call when an expired token accesses a protected endpoint

invalid_token_loader() is a function to call when an invalid token accesses a protected endpoint(for example, when cookie is missing)

After the token has expired/the user closed the browser and cookie has been deleted, if the user tries to reload the page or perform any actions leading to this(searching for restaurants, liking restaurants) it will be considered an unauthorised access attempt or access attempt with an expired token. Once it is detected, the user will be redirected to the login page to get access again