Adding JavaScript for signup route
This file receives values from the signup form after its submission. The password is validated and for each unfollowed rule (inclusion of uppercase letters/lowercase letters/digits/special characters) the user will receive an error message about the missing components. If the password structure has been validated, the credentials are received in the /signup route via POST method. If they have been varified there as well, the route function sends a response, which is caught in JS. Same happens when an error message is sent from the route function. According to this message, an alert appears in the top of the page, having a corresponding design and contaning the message from backend.

If registration has passed successfully, the user is redirected to the login page.

In line 32 the url is used for localhost and should be changed in production mode