Adding JavaScript for login route
JS for login route works simialar as for the signup route, yet does not perform validation. In case the response from the backend has been positive and the user has been varified, there is performed redirection to the home page. In case of errors with varifying the user, alerts with the corresponding messages are evoked.

Aagain,  the url in line 6 should be changed in production mode