Keeping user checks
User choices are now kepts in local storage, so after refreshing the page during the search, the checkboxes remain marked.
In case a refresh button is pressed, local storage is cleared, as well as filters and query string. In case the user logs out, local storage is cleared as well